chore: remove dead code
