+++ src/components/InfiniteScrollPaginator/index.ts
@@ -1,2 +1,111 @@
-export * from './InfiniteScroll';
-export * from './InfiniteScrollPaginator';
+import React, { useCallback, useEffect, useRef } from 'react';
+
+/**
+ * Prevents Chrome hangups
+ * See: https://stackoverflow.com/questions/47524205/random-high-content-download-time-in-chrome/47684257#47684257
+ */
+const mousewheelListener = (event: Event) => {
+  if (event instanceof WheelEvent && event.deltaY === 1) {
+    event.preventDefault();
+  }
+};
+
+export type InfiniteScrollProps = {
+  className?: string;
+  element?: React.ElementType;
+  hasMore?: boolean;
+  initialLoad?: boolean;
+  isLoading?: boolean;
+  listenToScroll?: (offset: number, reverseOffset: number, threshold: number) => void;
+  loader?: React.ReactNode;
+  loading?: React.ReactNode;
+  loadMore?: () => void;
+  pageStart?: number;
+  threshold?: number;
+  useCapture?: boolean;
+};
+
+export const InfiniteScroll: React.FC<InfiniteScrollProps> = (props) => {
+  const {
+    children,
+    element = 'div',
+    hasMore = false,
+    initialLoad = true,
+    isLoading = false,
+    listenToScroll,
+    loader,
+    loadMore,
+    threshold = 250,
+    useCapture = false,
+    ...elementProps
+  } = props;
+
+  const scrollComponent = useRef<HTMLElement>();
+
+  const scrollListener = useCallback(() => {
+    const element = scrollComponent.current;
+
+    if (!element || element.offsetParent === null) {
+      return;
+    }
+
+    // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+    const parentElement = element.parentElement!;
+
+    const offset = element.scrollHeight - parentElement.scrollTop - parentElement.clientHeight;
+    const reverseOffset = parentElement.scrollTop;
+
+    if (listenToScroll) {
+      listenToScroll(offset, reverseOffset, threshold);
+    }
+
+    // Here we make sure the element is visible as well as checking the offset
+    if (reverseOffset < Number(threshold) && typeof loadMore === 'function' && hasMore) {
+      loadMore();
+    }
+  }, [hasMore, threshold, listenToScroll, loadMore]);
+
+  useEffect(() => {
+    const scrollElement = scrollComponent.current?.parentNode;
+    if (isLoading || !scrollElement) {
+      return () => undefined;
+    }
+
+    scrollElement.addEventListener('scroll', scrollListener, useCapture);
+    scrollElement.addEventListener('resize', scrollListener, useCapture);
+
+    if (initialLoad) {
+      scrollListener();
+    }
+
+    return () => {
+      scrollElement.removeEventListener('scroll', scrollListener, useCapture);
+      scrollElement.removeEventListener('resize', scrollListener, useCapture);
+    };
+  }, [initialLoad, isLoading, scrollListener, useCapture]);
+
+  useEffect(() => {
+    const scrollElement = scrollComponent.current?.parentNode;
+    if (scrollElement) {
+      scrollElement.addEventListener('wheel', mousewheelListener, { passive: false });
+    }
+    return () => {
+      if (scrollElement) {
+        scrollElement.removeEventListener('wheel', mousewheelListener, useCapture);
+      }
+    };
+  }, [useCapture]);
+
+  const attributes = {
+    ...elementProps,
+    ref: (element: HTMLElement) => {
+      scrollComponent.current = element;
+    },
+  };
+
+  const childrenArray = [children];
+  if (isLoading && loader) {
+    childrenArray.unshift(loader);
+  }
+  return React.createElement(element, attributes, childrenArray);
+};
